Version 0.9 is out: - Added new flag for for turf/closed/wall/resin and /obj/structure/mineral_door/resin called "upgrading_now" for future use (like if you want to make sure wall/door is upgrading when animation play) - Returned old UI as option in designer ability wheel, now you can have 2 UI's active at same time. - Fixed Nodes deleting instantly when building at it, they should (i hope) delete -after- completion of construct. - Fixed Building on tile that contains speed node or cost node will now detect and apply buff correctly, even if you dont press on the node. - Fixed building on top of construct nodes, now you can build on top of construct nodes if designer die or forget to remove them. - Moved door opening/closing code to be proc(). - Now you can open/close doors with any ability selected in choose design (except old UI) - Removed "WEED_LEVEL_CONSTRUCT" from xeno.dm code, it was remains from older code. - Changed description of weedbound resin, so people dont mistake it for bein "weaker" (its just re-skinned drone wall).

+++ code/modules/mob/living/carbon/xenomorph/Powers.dm
@@ -103,10 +103,10 @@
 	if(!alien_weeds || alien_weeds.secreting)
 		return SECRETE_RESIN_FAIL
 
-	if(istype(target, /obj/effect/alien/resin/design/speed_node))
+	for(var/obj/effect/alien/resin/design/speed_node in current_turf.contents)
 		wait_time -= ((resin_construct.build_time * caste.build_time_mult) / 2)
 
-	if(istype(target, /obj/effect/alien/resin/design/cost_node))
+	for(var/obj/effect/alien/resin/design/cost_node in current_turf.contents)
 		total_resin_cost -= (total_resin_cost / 2)
 
 	var/obj/warning
@@ -151,8 +151,9 @@
 		playsound(loc, "alien_resin_build", 25)
 
 	var/atom/new_resin = resin_construct.build(current_turf, hivenumber, src)
-	if(istype(target, /obj/effect/alien/resin/design/speed_node) || istype(target, /obj/effect/alien/resin/design/cost_node) || istype(target, /obj/effect/alien/resin/design/construct_node))
-		qdel(target)
+	if(succeeded)
+		if(istype(target, /obj/effect/alien/resin/design/speed_node) || istype(target, /obj/effect/alien/resin/design/cost_node) || istype(target, /obj/effect/alien/resin/design/construct_node))
+			qdel(target)
 	if(resin_construct.max_per_xeno != RESIN_CONSTRUCTION_NO_MAX)
 		LAZYADD(built_structures[resin_construct.build_path], new_resin)
 		RegisterSignal(new_resin, COMSIG_PARENT_QDELETING, PROC_REF(remove_built_structure))
